mission changes

+++ Assets/Scripts/Missions/MissionTypes/AsteroidCollisionMission.cs
@@ -9,7 +9,7 @@ namespace StarSalvager.Missions
     {
         public BIT_TYPE? m_resourceType;
 
-        public AsteroidCollisionMission(BIT_TYPE? resourceType, string missionName, List<IMissionUnlockCheck> missionUnlockData, int amountNeeded) : base(missionName, amountNeeded, missionUnlockData)
+        public AsteroidCollisionMission(BIT_TYPE? resourceType, string missionName, List<IMissionUnlockCheck> missionUnlockData, float amountNeeded) : base(missionName, amountNeeded, missionUnlockData)
         {
             MissionEventType = MISSION_EVENT_TYPE.ASTEROID_COLLISION;
             m_resourceType = resourceType;
